refactor(console): move widget close to where it is used
