+++ gravitee-apim-console-webui/src/components/dashboard/widget/_widget.scss
