+++ gravitee-apim-console-webui/src/components/dashboard/widget/count/widget-chart-count.component.ts
