+++ gravitee-apim-console-webui/src/components/dashboard/widget/line/widget-chart-line-configuration.component.ts
@@ -15,7 +15,7 @@
  */
 import * as _ from 'lodash';
 
-import DashboardService from '../../../services/dashboard.service';
+import DashboardService from '../../../../services/dashboard.service';
 
 const WidgetChartLineConfigurationComponent: ng.IComponentOptions = {
   template: require('./widget-chart-line-configuration.html'),
