+++ gravitee-apim-console-webui/src/components/dashboard/widget/line/widget-chart-line-configuration.html
