+++ gravitee-apim-console-webui/src/components/dashboard/widget/map/widget-chart-map-configuration.component.ts
