+++ gravitee-apim-console-webui/src/components/dashboard/widget/pie/widget-chart-pie-configuration.component.ts
@@ -15,7 +15,7 @@
  */
 import * as _ from 'lodash';
 
-import DashboardService from '../../../services/dashboard.service';
+import DashboardService from '../../../../services/dashboard.service';
 
 const WidgetChartPieConfigurationComponent: ng.IComponentOptions = {
   template: require('./widget-chart-pie-configuration.html'),
