+++ gravitee-apim-console-webui/src/components/dashboard/widget/pie/widget-chart-pie.html
