+++ gravitee-apim-console-webui/src/components/dashboard/widget/stats/widget-data-stats-configuration.component.ts
@@ -17,7 +17,7 @@
 import { merge } from 'lodash';
 import { IOnInit } from 'angular';
 
-import DashboardService, { AverageableField } from '../../../services/dashboard.service';
+import DashboardService, { AverageableField } from '../../../../services/dashboard.service';
 
 interface Stat {
   key: string;
