+++ gravitee-apim-console-webui/src/components/dashboard/widget/stats/widget-data-stats-configuration.html
