+++ gravitee-apim-console-webui/src/components/dashboard/widget/table/widget-data-table-configuration.html
