+++ gravitee-apim-console-webui/src/components/dashboard/widget/widget.html
