+++ gravitee-apim-console-webui/src/index.scss
@@ -36,7 +36,7 @@
   @import 'components/documentation/error/error';
   @import 'components/alerts/quick-time-range/quick-time-range';
   @import 'components/user-autocomplete/user-autocomplete';
-  @import 'components/widget/widget';
+  @import 'components/dashboard/widget/widget';
   @import 'components/alerts/dashboard/alerts-dashboard';
   @import 'components/alerts/alert/alert';
   @import 'components/dialog/apiKeyMode/api-key-mode-choice.dialog';
